init private input
