+++ cli/src/lib.rs
@@ -1,3 +1,3 @@
+pub mod cli;
 pub mod commands;
 pub mod interactive;
-pub mod query;
