+++ examples/private-input/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "private-input"
+version.workspace = true
+edition.workspace = true
+license-file.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+exclude.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+hdp = { workspace = true }
+alloy = { workspace = true }
+tokio = { workspace = true }
