+++ hdp/src/primitives/mod.rs
@@ -2,6 +2,7 @@ pub mod aggregate_fn;
 pub mod block;
 pub mod merkle_tree;
 pub mod processed_types;
+pub mod request;
 pub mod serde;
 pub mod solidity_types;
 pub mod task;
