+++ hdp/src/primitives/processed_types/module.rs
@@ -1,11 +1,9 @@
 use alloy::primitives::{Bytes, B256};
 use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
-use starknet::core::serde::unsigned_field_element::UfeHex;
-use starknet_crypto::FieldElement;
 
-#[serde_as]
+use crate::primitives::task::module::ModuleInput;
+
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct ProcessedModule {
     /// encoded computational task
@@ -15,8 +13,7 @@ pub struct ProcessedModule {
     pub task_proof: Vec<B256>,
     pub result_proof: Vec<B256>,
 
-    #[serde_as(as = "Vec<UfeHex>")]
-    pub inputs: Vec<FieldElement>,
+    pub inputs: Vec<ModuleInput>,
     /// Detail class code of the module.
     /// This will be loaded to bootloader.
     pub module_class: CasmContractClass,
@@ -29,7 +26,7 @@ impl ProcessedModule {
         result_commitment: B256,
         task_proof: Vec<B256>,
         result_proof: Vec<B256>,
-        inputs: Vec<FieldElement>,
+        inputs: Vec<ModuleInput>,
         module_class: CasmContractClass,
     ) -> Self {
         ProcessedModule {
