+++ hdp/src/primitives/request.rs
@@ -1,4 +1,4 @@
-use hdp::primitives::task::{datalake::DatalakeCompute, module::Module};
+use crate::primitives::task::{datalake::DatalakeCompute, module::Module};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -72,16 +72,26 @@ fn test_serialize_submit_batch_query_datalake() {
 #[test]
 fn test_serialize_submit_batch_query_module() {
     let json_data = r#"
-    {
-      "destinationChainId": 11155111,
-      "tasks": [
+   {
+    "destinationChainId": 11155111,
+    "tasks": [
         {
-            "type": "Module",
-            "programHash": "0xaf1333b8346c1ac941efe380f3122a71c1f7cbad19301543712e74f765bfca",
-            "inputs": ["0x4F21E5", "0x4F21E8", "0x13cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"]
-        }        
-      ]
+        "type": "Module",
+        "programHash": "0x64041a339b1edd10de83cf031cfa938645450f971d2527c90d4c2ce68d7d412",
+        "inputs": [
+            {
+            "visibility": "private",
+            "value": "0x5222a4"
+            },
+            {
+            "visibility": "public",
+            "value": "0x00000000000000000000000013cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"
+            }
+        ]
+        }
+    ]
     }
+      
     "#;
 
     let parsed: SubmitBatchQuery = serde_json::from_str(json_data).unwrap();
