+++ request.json
@@ -5,8 +5,14 @@
       "type": "Module",
       "programHash": "0x64041a339b1edd10de83cf031cfa938645450f971d2527c90d4c2ce68d7d412",
       "inputs": [
-        "0x5222a4",
-        "0x00000000000000000000000013cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"
+        {
+          "visibility": "private",
+          "value": "0x5222a4"
+        },
+        {
+          "visibility": "public",
+          "value": "0x00000000000000000000000013cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"
+        }
       ]
     }
   ]
